backend/pyproject.toml
[project]
name = "grimdm"
version = "0.1.0"
description = "GrimDM - Voice AI Dungeon Master powered by Gemini Live API"
readme = "../README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "ScrappyLabs" }
]

dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
firestore = [
    "google-cloud-firestore>=2.19.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
